updated expenditure flow

+++ app/api/expenditures/[id]/route.ts
@@ -1,15 +1,26 @@
 import { NextRequest, NextResponse } from "next/server";
 import { prisma } from "@/prisma/config";
+import { verifyOrgRole } from "@/lib/auth/checkRole";
+import { verifyBusiness } from "@/lib/auth/checkBusiness";
 
 export async function PUT(
   request: NextRequest,
   context: { params: { id: string } }
 ) {
   try {
-    const { amount, description, date } = await request.json();
-    if (!amount || !description) {
+    const { amount, description, date, userId, businessId } =
+      await request.json();
+
+    if (!amount || !description || !userId || !businessId) {
       return NextResponse.json({ error: "Missing fields" }, { status: 400 });
     }
+
+    const businessUnauthorized = await verifyBusiness(userId, businessId);
+    if (businessUnauthorized) return businessUnauthorized;
+
+    const unauthorized = await verifyOrgRole(userId);
+    if (unauthorized) return unauthorized;
+
     const expenditure = await prisma.expenditure.update({
       where: { id: context.params.id },
       data: { amount, description, date: date ? new Date(date) : undefined },
@@ -28,6 +39,18 @@ export async function DELETE(
   context: { params: { id: string } }
 ) {
   try {
+    const { userId, businessId } = await request.json();
+
+    if (!userId || !businessId) {
+      return NextResponse.json({ error: "Missing fields" }, { status: 400 });
+    }
+
+    const businessUnauthorized = await verifyBusiness(userId, businessId);
+    if (businessUnauthorized) return businessUnauthorized;
+
+    const unauthorized = await verifyOrgRole(userId);
+    if (unauthorized) return unauthorized;
+
     await prisma.expenditure.delete({ where: { id: context.params.id } });
     return NextResponse.json({ success: true });
   } catch (error) {
